Cache isthmuses, total number of plots
[change id advc.opt]

The presence of isthmuses blocking sea movement is the only check that the pathfinders have to make on a per-edge basis (as opposed to per node). Need only 1 bit to pre-compute and store whether there is any isthmus adjacent to a water plot. (For the vast majority of water plots, there is none.)

The total number of plots was constantly being re-computed by CvMap as grid width times grid height.