BUG option for disabling savemap key combination
Mainly to make the savemap feature more visible. But it's also conceivable that some player keeps pressing those keys by accident and therefore wants to disable them. [advc.savem]

Savemap is now always disabled in network games because it could be used as a maphack cheat. [advc.savem]

And I've disabled the "auto moves" feature of AI Auto Play because it's misleading -- 1 turn of AI Auto Play is not the same as just letting the AI move all human units -- and because the key combination Ctrl+Shift+M is too similar to the savemap combination Alt+Shift+M. [advc.127]