BUG option for total output next to culture slider
[change id advc.004p]

Related CFC post:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-39#post-15843459

And removed one of the three dropdown menus for customizing the color of the gold rate. A negative gold rate that exceeds the available funds is now shown in the same color as any other negative gold rate. Because the menu was getting crammed and the gold rate color is very unimportant. [advc.070]