Fix AI not respecting AdvCiv rule about tech requirements
For example, the Earth1000AD scenario lets India start with Paper but not with any of the prerequisite techs of Paper. In this case, India should not be allowed to research Education (which requires Paper) until India has discovered the requirements for Paper. The AI had been ignoring that rule (since the update to K-Mod 1.46 I think). [advc.126]