UWAI: Tweak "loathing" aspect of war utility calculation
Disregard the loathing aspect when expecting to be eliminated soon. This should prevent the AI from continuing suicidal wars just to inconvenience an enemy.

When computing war utility from the point of view of a human civ (relevant e.g. for peace terms), don't completely disregard the loathing aspect, i.e. assume that human players are to an extent vengeful.