UWAI: Fix some issues with AI capitulation decisions
These changes should align the "not right now" response (computed quickly on the human turn by CvTeamAI) better with the decisions made by the UWAI component (on the AI turn).

And, in situations when an AI civ is fighting multiple wars, the all-war-ending effect of a capitulation is now taken into account. (Not reliably, but I hope it'll work well enough most of the time.)

Also fixes a flipped sign (wasn't terribly consequential though).

To address problems apparent from the savegames posted here:
https://forums.civfanatics.com/threads/quick-advciv-questions-answers.666137/#post-15997844