Show best visible enemy interceptor in tile hover
When hovering for Air Strike, Air Bomb or Paradrop. When a human group is ordered to undertake one of these missions, unless Stack Combat is enabled, the execution stops when an interception occurs, giving the player a chance to reconsider his or her orders. [advc.004c]

And revised my earlier recent changes to the Air Bomb UI:
66373d4b04f732c8bfa665795182c527488c822e