Use only a single BUG option for "top cities" icons
The individual options are still operational on the DLL-side, just commented out in XML and Python. The single option does not cover the icon for the Globe Theater and National Park cities; those two (recently added) icons are, for the time being, entirely disabled. The icon for the city with the highest free XP is no longer shown for a city with just a Barracks. [advc.002f]