Stricter conditions for AI opening borders for naval trade
A revealed land tile now only makes an AI civ willing to open its borders if the tile is on the same landmass as a city owned by the AI civ. If it's on a different landmass, then there needs to be a trade connection (via Sailing or Astronomy). The AI will also open its borders if it has a unit adjacent to a tile that it can't currently enter but will be able to enter through an OB agreement. (And, as before, the AI is willing to open its borders regardless of accessibility if its attitude level is one above the threshold set in XML.) [advc.124]