Wider civics screen
Only intended as an adaptation to the larger fonts in AdvCiv. [advc.002b]
The panels on the bottom of the screen were a bit too narrow. I've left the ones on the top alone; they look better to me without extra space. Also, I'm not letting the civics screen take up the whole available width because that much space isn't needed; better to leave room for the scoreboard instead I think.