Refactor combat odds calc, damage limit bugfix
Move the getCombatOdds and LFBgetCombatOdds functions into a new translation unit, split them up so that they can share some code with each other and so that the Advanced Combat Odds mod (ACO) can also reuse some of it. (To be done in another commit: Throw out the code now in calculateCombatOddsLegacy, move the ACO code.) Refactored all the odds-related code, but mostly just the variable names (which had made the ACO code quite difficult to comprehend).

Primarily, this should make it more feasible for mod-mods to change the combat rules. (Updating the odds calculations is a real obstacle to that.)

Added code for verifying the calculated odds through simulations of the combat resolution procedure. Found one bug this way, which had already been fixed in ACO (i.e. the displayed odds were wrong only when ACO was disabled): The BtS/ Lead From Behind odds had assumed that combat ends when the attacker's damage limit is reached, but combat resolution actually requires the limit to be exceeded. This matters for example when a Cannon attacks another Cannon because the damage limit (80) is then a multiple of the per-round damage.

I've chosen to fix the bug by changing the combat resolution procedure -- combat now ends upon reaching the combat limit. I've updated the ACO code accordingly. The hit that reaches the damage limit is now also included in the Combat Log. [advc.001l]