Revised all code in the UWAI component
... and also AdvCiv4lerts.

I've tested this a bit; looks like there can't be many new bugs left. Will have to do a side-by-side review of the changes later (and do a lot more testing); for now I just want to have it committed.

The style now conforms with the style that I use in the rest of the GameCore DLL. Floating-point numbers been replaced almost entirely with fixed-point numbers (ScaledNum), in the UWAI component and in all other synchronized code, the exception being a couple of uses of std::log and the (BtS/ Lead From Behind) combat odds calculations. if nothing else, this seems to have improved performance - though other factors could be responsible for that as well. A lot of low-level refactoring, some bugfixes; the important ones, I'll commit separately on another branch.