Make sure that normalization upgrades Desert to Plains
In the code, it looks like it had upgraded to Grassland before, and I've seen Oil on Grassland on a PerfectMongoose map. That said, I've only ever seen Incense on Plains, never on Grassland. Not sure what the problem is here, if any. Let's make sure and clear that it should be Plains and doesn't hinge on the order of terrain types in XML. [advc.108]