Fix GameFont issue (amendment)
Amends
b40050a135c64e8086882827c0f4c7883ba388b7

It seems that the particular number of symbols and assignment to cells in that commit prevents the friendly-attitude smiley from being loaded. I hear that the GameFont file can be weird like that. I've now thrown out a duplicate of the cautious-attitude smiley -- which, I think, was intended as a generic attitude icon and, as far as I can tell, was unused. [advc.187]