EnumMaps for nested XML data
And a macro for iterating over the (non-default) values in those maps. [advc.003t]
Used only in CvInfo_Building so far (one of the two classes with a lot of nested data, the other being CvInfo_Unit).
CvInfoMap2D will be removed in the next commit; works but is not needed.

Inspired by the InfoArray class in the "We the People" (WtP) mod:
https://forums.civfanatics.com/threads/implemented-major-upgrade-to-building-construction-rules.670368/
(But I haven't adopted any of the WtP code.)