Merge AI_nukeRange into AI_nuke
Both had been rewritten by K-Mod. The one was for Tactical Nukes, the other for ICBMs. There shouldn't be fundamental differences. The AI may now also target stacks outside of cities with its ICBMs. (The AI is still terrible at this, however, because it doesn't realize that dropping a single nuke on a stack is generally pointless.) [advc.650]