SparseEnumMap class, mainly for per-civ data at CvPlot
Scarcely makes a difference with an 18-civ limit, but seems to speed up AI turns by some 15% with a 31-civ limit. The penalty for allowing 31 civs while playing only with 18 is now down to less than 10% on Earth18 (super-Huge map) in a test, and should hopefully be less than 5% on regular Huge maps (was 8% in a test prior to this change). [advc.enum]