Halve AI deal cancellation memory when AI cancels
That is, when Open Borders or Defensive Pact is canceled by one side, the other side (if it's an AI civ) refuses to renew the deal for, on average, 10 turns (no change) and the side that canceled the deal (if it's an AI civ) refuses a new deal for 5 turns on average. [advc.130j]

Should make it a bit less frustrating when the AI cancels OB due to AI diplo attitude dipping below the OB threshold.