Remove most inline keywords
[change id advc.inl]

The keywords are nice for making the reader aware that a function is very frequently called, but I don't think it's ultimately worth the clutter, and as for performance ... Maybe I'll compare once I've put together this commit. I'm pretty sure that there's not much of a difference, but I've made other changes along the way, so it's difficult to say right now.

I've also removed inline keywords that (I think) standard C++ would require for the one-definition rule, but that are dispensable in MSVC03. Seems to be that function templates and members of class templates never need an inline keyword.

Where "inline" is not required by the linker, "__inline" is now preferred; for clarity.

For free functions, I still use __inline a lot so that even debug builds - through the Ob1 compiler option - can inline them. (Apparently, the Ob1 option can always inline member functions defined in a class definition, but, elsewhere, an inline keyword is required. That's also what the VS documentation says.)

And I'm removing some of the "...External" counterparts of inline functions. (The ones I'm keeping exist because of some signature change; not to avoid exported inline functions.)