Cache plot index at CvPlot
Letting CvMap compute the index takes a multiplication and an addition (multiplicand and addend being unknown at compile time), and it's unnecessarily verbose when the CvPlot instance is already at hand. Should be worth the 2 extra byte (which were taken up by padding anyway). [advc.opt]

And revert a list-based enum map back to an array-based one: access to revealed routes needs to be fast for pathfinding. [advc.enum]