Group pathfinder prefers landing on routes
[change id advc.pf]

When breaking ties between paths of equal length, paths that end on a tile with a route are now preferred. And AI units now have the same preference (i.e. only a slight one) for diagonal moves as human players. This gives other tiebreakers (including the one based on routes) greater importance.

Apart from yielding, in my estimation, slightly better paths, this change also works around a rare issue that can cause valid paths to be ruled out when avoiding dangerous tiles. The specific situation seems worth documenting:

I've gotten an AI worker trying to retreat toward a city stuck in a loop because the clean pathfinder instance in CvSelectionGroup::groupPathTo did not find any valid path whereas the pathfinder call in CvUnitAI::AI_retreatToCity did find one with the help of data cached from previous calls. ASCII art:
| | | |D
|A|B|C|
|H|E| |
A is the start, D the destination, and B, C and D have a Road; A and E do not. H contains a hostile unit. Due to the diagonal bias (and no bias for moving onto a route), the pathfinder first processed E, and found that entering C from E is not allowed because the worker would then end the turn on a tile threatened by H. This causes the pathfinder to permanently regard C as a dead end, without consideration of the cheaper path via B that allows the worker to reach the safe tile D in a single turn.

It's not correct to let the canStepThrough function ("pathValid_source" in K-Mod) take path data (remaining moves) into consideration. Strictly speaking, danger can't be checked by that function. I guess it could be done in isValidStep ("pathValid_join" in K-Mod), but I expect that this would hurt performance significantly. Not worth it for such a rare problem -- that will become even more rare (impossible?) through my change to the tiebreakers.

This example suggests to me that the route bias may, more generally, help the pathfinder process promising paths first. That might lead to a small gain in performance.