Fix AdvCiv bug in tech cost calculation
Negative cost modifiers from Civ4EraInfos.xml and from the Raging Barbarians option hadn't been applied. This bug was introduced in v0.97, specifically by this commit:
82b11c4893afec5e100af6eb0722cc91e700bea7

The pace of research has been fine in recent tests, so I'm setting the negative modifiers (for the Classical and Medieval era) to 0. [advc.910]
So, on the bottom line, the bug hadn't really been a problem. (The tech cost adjustment from Raging Barbarians [advc.308] is very minor.)