Prevent AI-AI resource trades that aren't win-win
I seem to have missed that such a condition needs to be checked when I implemented the more rational AI for resource trades. [advc.036]

The condition that I did put in place in commit
23137d203546828371fc81e67d132ef297c78014
wasn't adequate either. In particular, the AI was (at times) still trading away crucial strategic resources.