Teleported AI units don't always lose moves
Generally, forced moves shouldn't be free. I'm making an exception for AI units that share a tile with an enemy unit after war is declared. Because it has occurred to me that human players could otherwise delay AI invasions by placing e.g. a worker in the path of the approaching AI stack. [advc.163]