Reduce shared-religion diplo for half of the leaders
Namely for all leaders with a SameReligionAttitudeChangeLimit of 5 or greater (this also includes all leaders with a favorite civic that implies a state religion) and for about half of those with a limit of 4. I've reduced the limits by 1 point. [advc.148]

Note that this limit does not include the flat (not time-based) +1 relations from SameReligionAttitudeChange, nor the +1 relations from the Holy City owner.

Related CFC post (1st spoiler box):
https://forums.civfanatics.com/threads/advanced-civ.614217/page-50#post-16166874