Reduce shared-civic diplo for half of the leaders
Namely for everyone with a FavoriteCivicAttitudeChangeLimit of 4 or greater except Lincoln, Mansa Musa, Washington and Willem. By exempting those four (which all have a limit of 4), I end up reducing the limit of almost exactly half of the AI leaders. The limits get reduced by 1 point. [advc.148]

Note that the limit does not include the flat (not time-based) +1 relations from FavoriteCivicAttitudeChange.