Streamline religion, civic AI diplo adjustments
The different-religion penalty now starts at -1 (or -2 with the Holy City) for all AI leaders. For the majority of the leaders, the penalty then becomes one worse after 15 turns and, for some of the leaders, another one worse after another 15 turns. The total limits remain the same for all leaders. This is the same principle as for the same-religion and same-civic modifiers (but those increase every 10 turns and, overall, can grow bigger than the different-religion penalties).

Previously, different-religion penalties had started at 0 to -2 (another -1 with the Holy City) and had then grown one worse after 5 turns.

Now that all three ideological modifiers (same/ different religion, same civic) have a small constant portion and a larger time-based portion, the time-based portions are all essentially given the same adjustment to the popularity of the ideology, specifically the number of cities of known civs with the ideology in question. [advc.130n]

This obsoletes change [advc.130x] which had adjusted the different-religion penalty based on the population sizes of known cities with the two religions in question. Change id 130x instead refers to the following accompanying change now: Each time that a time-based ideological modifier increases, the time for the next increase is incremented by one turn.

This commit also takes care of attitude cache updates after loading older savegames -- which are also needed for additions to change [advc.148] previously committed.

Related CFC post (1st spoiler box):
https://forums.civfanatics.com/threads/advanced-civ.614217/page-50#post-16165134