Fix UWAI bugs that had made the AI too peaceable
Two of these were introduced by the stylistic overhaul in v1.00:
9814ada3407d4b4f892aa7571bbcec103b96bbd9

A variable had not been reset in between evaluating the diplomatic ill will from a potential war enemy and its partners. This had caused the costs for ill will to be counted multiple times. Except in situations when relations with the target were already spoiled, this had made the AI far more reluctant to start wars than it should've been.

Utility for reparations had been counted even when neither side was expected to lose cities, in particular also in the peace scenario. Since the total utility of a war plan is the difference between the war and peace scenario, possible reparations had effectively not been taken into account and war utility was a bit lower than it should have been in many situations.

The last bug was that UWAI had ignored the time horizon of a simulated war when estimating how much the military power of civs not involved in the war will increase. It had instead always done a 10-turn projection and thus underestimated the military power of those civs and therefore also underestimated the post-war utility of units produced for a war. On the bottom line, militarily powerful civs had tended to (not dramatically, but significantly) overestimate the opportunity cost of starting a war. I think this oversight had existed since the initial release of the mod.