Make UWAI slightly more ready to entertain war trades
The utility threshold for "enough on our hands" exists mainly so that human players aren't inundated with alerts about unattractive war trades. And so that the AI doesn't needlessly signal its intentions. I haven't been seeing a lot of war trades offered, so I'm relaxing the threshold a little. (That said, a better way to get more war trades would be to relax the BtS attitude thresholds.)