Make UWAI more pessimistic about possible tech trades
This change makes the AI a little less reluctant to hurt (or sever) relations with a civ with whom tech trades might be possible in the future. Yet another tweak to encourage more warfare in the Classical era.

And the game options for No Tech Trading and Brokering are now taken into account.