AI: Don't adopt Space3/4 strat without any SS parts
Ensure that a substantial portion of the needed Spaceship parts have been produced before switching to stage 3 or 4 of the Space victory strategy. BBAI/K-Mod had only required sufficient tech progress. Important not to switch to stage 3 or 4 of any victory strategy too early because UWAI looks at those stages to determine which rival civs are on track to victory. [advc.115]
In a much older change, I had already addressed a similar issue for stages 1 and 2 (AI shouldn't adopt those when the available production capacity is clearly insufficient).