Tweak Barbarian activity peak
So that activity increases, peaks a bit earlier. I've tweaked this before, with the mod's increased default player counts in mind. By now, I've also come to think that it's more fun to have some more pressure from Barbarians around the time that the 2nd and 3rd city are founded. And one should also bear in mind that players accustomed to the earlier activity peak in BtS may find the Barbarians very tame early on in the mod - not an ideal. [advc.300]