Don't sanitize starts of players with fixed civ
I think it's cleaner to fully exempt players with a non-Random civ from the True Starts code. The way it's implemented now, a bunch of unnecessary calculations are also skipped for the fixed-civ players.

And fixed civs are no longer ignored for all scenarios; just those that don't even allow Random civs.