Remove Archery req. from Great Wall
Now that Barbarian activity peaks earlier again, I think the Great Wall needs to get a little easier to get too, especially for the Raging Barbarians option. I didn't want to reduce the cost (because it's an extremely massive construction). While it's nice to give Archery another purpose, the historical explanation was a bit strenuous, and one important purpose of the Archer unit is to fend off Barbarians, which conflicts with the Great Wall effect. [advc.310]