AI workers avoid tiles threatened by potential enemies
Specifically, when deciding which tile to improve, tiles credibly threatened by a rival human unit are deprioritized.
Through change advc.010, workers can't easily be captured by surprise, but they can still be easily destroyed, which is a powerful attrition tactic. So it would be nice to get the AI to anticipate such attacks. However ...
Open issues:
- A worker that chooses to improve a safe tile may still move through a threatened tile along the way.
- The AI doesn't move workers to safety that become threatened while already building an improvement. Code for making such workers consider a different tile was easy enough to write, but I've commented it out because the AI will frequently move through another threatened tile along the way. Moreover, my code doesn't handle back-and-forth moves of threatening units well.
- Better results could be achieved by coordinating with nearby military units. Related CFC post:
https://forums.civfanatics.com/threads/k-mod-far-beyond-the-sword.407049/page-147#post-15164375

This would need a fair amount of additional work, in particular, a new pathfinder flag and some persistent choice about which tile to improve when the threat keeps moving. It's doable, but I don't want to spend entire weekends on this. And the AI will still be at a disadvantage because preferring safe improvements has an economical cost. Human players would incur the same cost only if the AI were also capable of harassing human workers.

So I've no further plans for this branch.