Let Tectonics place fewer hills and peaks
Now it's sometimes as few as on most scripts, but usually still more. Placement is based on (absolute) elevation, so I guess the script needs to place lots of hills or else large regions will end up without any. I've randomized the placement a bit to alleviate that problem. [advc.021a]