BUG option for HUD scaling; finish unit plot list scaling.
Apart from the BUG option, I've also implemented formulas for scaling the different types of distances (widget side lengths, button sizes, spaces) in vertical and horizontal direction. Results so far look good - except for all the small widgets that I haven't yet gotten too.

One tricky issue with the frame around the map excerpt on the city screen became apparent: the panel that forms the top of that frame can't be resized vertically. I've solved this (seamlessly I think) by additionally showing an overlapping duplicate of that panel at a vertical offset.

Regarding the plot list, it was mostly the BUG drawing method that still had to be taken care of. I've committed a related change to the 1.06 branch: support for the BUG drawing method when shifting the plot list through the arrow buttons.

When using the PLE drawing method, right-click on an arrow button will now scroll the plot list all the way in one direction.

(Also on the 1.06 branch: City screen camera position adjusted so that the excerpt of the map is actually centered on the city tile. That should alleviate problems with small field-of-view values causing the enlarged HUD to cover of part of the city's working radius.)