BUG option for showing building icons on city screen
And the names of obsolete buildings are grayed out. This part is not optional. I've also written some code for sorting the building list, but wasn't happy with my ordering and left it commented out. [advc.097]
Merged "Show Only Present Religions/ Corporations" into a single option. [advc.004]
Resource amount on the city screen are shown in yellow, as on the Foreign Advisor. [advc.092]

Scaling of the building list layout is mostly still to be done.