Fix BBAI bug in PathToEnemyCity check
Originally a bug in BBAI that had survived the AdvCiv rewrite of that function. Thanks to spqkfk for making me aware. The bug had caused coastal AI cities to produce transports while preparing war against a land target that the city owner did not have Open Borders with. [advc.001]

Also, needlessly, added a cache for some text displayed in Debug mode. It bothered me to think that the paths are constantly recalculated while the text is being displayed.