Partly restore BtS behavior for choice of "center unit"
Units with available moves are - again - preferred as the center unit of a tile, i.e. the unit that gets shown on the map and - as I've realized through this CFC post
https://forums.civfanatics.com/threads/advanced-civ.614217/page-52#post-16278294
- the unit that gets selected when a tile is clicked.

I had disabled the available-moves condition as part of a bugfix: karadoc had intended to disable that condition, according to a comment, but he had accidentally disabled a different one instead.

Now available moves are checked again, but only during the active (human) player's turn. That should be the best of both worlds, but could also be confusing when a center unit changes upon ending a turn. Let's see.

I've also refactored the calls to getBestDefender to avoid passing half a dozen bool parameters.