Reorder unit command buttons, use two rows.
[change id advc.004k]

More or less swapped the position of misc. missions and commands (but the important ones have an "order priority set explicitly through XML) and moved promotions and upgrades to the very (right) end and, if there is space, into a separate row.

Put Heal before Skip, Air Patrol a bit farther to the right, Sea Patrol, Gift and Delete much farther to the right. At the end of the first row, there is now normally either a split-group command or a Cancel command and before that Delete - which used to be all the way to the left. The result is that rarely used and dangerous buttons are now near the right end, but not all the way.

No option for this because it's in part implemented in XML and through the order of the CommandTypes enum.

Related CFC post (under "Unit command buttons"):
https://forums.civfanatics.com/threads/advanced-civ.614217/page-53#post-16292755