Move found-city mission button all the way to the right
The reordering of unit action buttons had put it to the left of Delete. That feels quite unfamiliar right at the start of a game. I don't think Delete being to the left of a mission button in this one instance is going to bother players; the distinction between missions, commands, automation actions, grouping buttons etc. is mostly an internal technical matter. [advc.004k]