Fix a couple of minor bugs in PerfectMongoose
One of the bugfixes, in code for breaking up clusters of peaks, is from Totestra - but I've decided to disable that behavior entirely, which, I think, had been the effect of the bug as well, so this change is really just clean-up. The other bug concerns the number of "meteors" for breaking up a continent. [advc.021b, advc.001]