Function for misc. game speed adjustments
Already had one, but it's better not to use ScaledNum for this, and the old function had only two call locations. The new function (CvGame::getSpeedPercent) now replaces all uses of CvGameSpeedInfo::getVictoryDelayPercent and getGrowthPercent that don't actually deal with the Spaceship or population growth. [advc.252]

As the only functional change, two uses of getVictoryDelayPercent in CvTeamAI::AI_doWar (checks for AI-AI war having gone on for too long) have been replaced with getTrainPercent. Only relevant when UWAI is disabled.