Don't show red waypoint for attack when out of moves
More specifically, when a unit has already attacked, BtS will indicate that further attacks are illegal through a red waypoint. This is misleading when the target of the attack can't be reached on the current turn - because, by the time that the attacker gets there, it will again be able to attack. My fix only addresses the (typical) case in which the attacker has spent all its moves. (Because this case is simple - no target can then be reached on the current turn, so no red waypoint should be shown for any target. [advc.004]