Fix unit cycling getting stuck when some units have no moves
Now treated the same way as when all selected units have no moves - queue the mission, cycle to to the next group. [advc.001]

The modified code is from K-Mod, but this issue had already existed in BtS.

Not sure that my change won't have unforeseen side-effects. Will have to wait and see.