Show city trade help on Diplo screen
Hover text for cities (which is supposed to explain what the effect of a city trade is) had previously just shown "Cities". Now shows the same text as the Cities tab on the Foreign Advisor. [advc.ctr]
(Which still isn't much - would be nice to show, roughly, how much the city is worth to the AI.)