Fix AI issues caused by disabled Culture victory
Inspired by this CFC post:
https://forums.civfanatics.com/threads/realism-invictus.411799/post-16511367
I don't think this particular problem had existed in AdvCiv/ K-Mod, but it almost did - precarious.
And it seems that there was a small K-Mod issue with AI cities trying to reserve high-culture buildings for the top culture cities even when culture victory was disabled. The AI now treats the city threshold for Culture victory as a very high number when Culture victory is disabled - instead of treating that threshold as 0.