Make AI_isControlled nonvirtual (for speed)
I had kept a single CvSelectionGroup function pure virtual so that the class is abstract, which still seems reasonable to me, but I shouldn't have picked such a frequently called function. Rather than add a lot of (wrapper-enclosed) casts to CvSelectionGroupAI, I've moved the function to the base class, where it imo makes more sense anyway. And "AI_isControlled" was a silly name; now "isAIControlled". [advc.003u]

Doing this also with the latest pathfinder tweak in mind; I think that will make AI-control checks a lot more frequent. [advc.pf]

With both changes in place, the performance seems to be essentially unchanged based on a few measurements I've run. Which is fine with me.