Initial attempt at adjustable help text area width
Still needs some work to make it safer. [advc.092c]

CFC thread to be credited:
https://forums.civfanatics.com/threads/modifying-plot-help-width.693581/

The FONT_SIZE_FACTOR (14) becoming HELP_FONT_SIZE (16) could be troublesome for mod-mods; they usually have their own changes in GlobalDefines_advc.xml and don't always think of updating that file when I change it. Changing the name of the variable should actually help because, if the new name doesn't exist, the DLL falls back on a default value of 16 and probably all will be well.