Allow font size to be adjusted from GlobalDefines
Unsure if this will go on the main branch. [advc.002b]
Advantages: A more visible, user-friendly font size setting. The mod can work from a My Games path. The mod folder can easily be renamed.

Disadvantages: Renaming the mod folder remains somewhat inadvisable (and is no longer warned against) because it'll cause a new .ini file to be created in the mod's folder, which will, by default, allow (all) CustomAssets to be loaded alongside AdvCiv - and because a variety of AdvCiv mod folder names will cause those versions to have mutually incompatible savegames.
The size adjustments are implemented through a hack of the EXE which might turn out not to work with all editions of the game. That perhaps can be straightened out, but temporary problems are bad enough.
The use of the original GFC theme files will cause ThemeParseLog.txt with error messages to be created in the BtS install folder due to (minor) errors in those files. Using modified versions allows those errors to be corrected. (Granted, players will still get a ThemeParseLog.txt with errors unless they play exclusively with AdvCiv - I don't think any other mods fix those errors.)
May cause users who have already adjusted the font sizes (or other UI settings) through the GFC theme some trouble when updating.

This change would be a clearer improvement if font sizes could be adjusted from the BUG menu. This could probably be done (see the branch early-BUG-options), but a restart of Civ would still be necessary for size changes to take effect. For the technical details of the implementation, see this CFC post:
https://forums.civfanatics.com/threads/modifying-plot-help-width.693581/#post-16731936