Adopt yield icons from Realism Invictus
From some old version, but I suppose they still look that way in the current version. Including them in GameFont.tga like this won't affect the yield display in the center of the city screen (and optionally on the main map). I think I'm OK with that. Piled up like that, the new icons look rather alienating, but I like them in game text and I hope they're similar enough to the original icons that the mix of the two versions won't be distracting. [advc.002r]

If I do want the RI icons on the map, I think the files to adopt are  in Art\Interface\Symbols. Copying the new icons into the old mip-mapped vertical_symbols.dds atlas probably won't work well. Better to take the RI version of that file along with the symbols NIFs. Possible that some Art\XML paths also need to be adjusted. But I think I'd sooner revert to only the old icons.

Mod-mods with their own GameFont.tga imo shouldn't bother to adopt this change. Fwiw, I've used an old Photoshop version to edit the TGAs, making sure that the alpha channel is selected in the Channels window.