Tweak AI eval. of contested city sites
"Contested" mostly in the sense of deserving to be prioritized because a neighbor might beat the AI to it. But this bit of (K-Mod) code also deals with possible cultural domination when settling too close to a rival and too far from one's own cities. This latter part now is especially wary of rival capitals. And, when looking for a rival that might be interested in the same spot, I've extended the range a little. [advc.031]
Related CFC post (after 1st quote box):
https://forums.civfanatics.com/threads/advanced-civ.614217/post-16754529