Update pathfinder waypoints after signing OB
Also in a couple of other situations in which the active player may gain more freedom of movement. [advc.001w]

I had previously addressed a similar problem after declaring war in this commit:
61ca92bbd0a8d42c4a9168701698c04d6bf12176

Not sure what the line tagged with advc.162 was all about. I think it's probably needed if that change is enabled. But doing it only for the active player looks like it could've caused network sync problems.

Bug report regarding OB:
https://forums.civfanatics.com/threads/ai-city-placement-and-misc-suggestions.695343/page-5#post-16778254